+++ bootstrap/setup-php.sh
@@ -13,16 +13,16 @@ try
 (
 	throwErrors
 
-	echo "Add 'ondrej' repository to Aptitude"
+	echo "Add 'sury' repository to Aptitude"
 	wget --quiet -O - https://packages.sury.org/php/apt.gpg | apt-key add -
-	echo "deb https://packages.sury.org/php jessie main" >/etc/apt/sources.list.d/ondrej.list
-	echo "deb-src https://packages.sury.org/php jessie main" >>/etc/apt/sources.list.d/ondrej.list
+	echo "deb https://packages.sury.org/php jessie main" >/etc/apt/sources.list.d/sury.list
+	echo "deb-src https://packages.sury.org/php jessie main" >>/etc/apt/sources.list.d/sury.list
 	apt-get -y -q update
 
 	echo "Install packages"
 	apt-get -y -q install php7.1 php7.1-cli php7.1-common php7.1-curl php7.1-dev php7.1-fpm php7.1-gd \
 	                      php7.1-intl php7.1-json php7.1-mbstring php7.1-opcache php7.1-pgsql php7.1-sqlite3 \
-	                      php7.1-xdebug php7.1-xml php7.1-zip
+	                      php7.1-xml php7.1-zip php-xdebug
 
 	echo "Disable OPcache"
 	sed -i 's/;\?opcache.enable=.\+/opcache.enable=0/g' /etc/php/7.1/fpm/php.ini